pyproject.toml
[tool.poetry]
name = "detectron2-effizency"
version = "0.1.0"
description = ""
authors = ["ranhomri <ran.homri@tensorleap.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9, <3.12"
numpy = "*"
setuptools = "*"
mock = "*"
fvcore = "*"
matplotlib = "*"
opencv-python = {extras = ["headless"], version = "^4.8.1.78"}
torch = "*"
torchvision = "*"
Pillow = "*"
pycocotools = "*"
yacs = "*"
termcolor = "*"
cityscapesscripts = "*"
tabulate = "*"
tqdm = "*"
scipy = "*"
lvis = "*"
psutil = "*"
onnx = "*"
python-panopticon = "*"
black = "*"
isort = "*"
av = "*"
iopath = "*"
omegaconf = "*"
PyYAML = "*"
submitit = "*"
cloudpickle = "*"
packaging = "*"
timm = "*"
pandas = "*"
fairscale = "*"
pytorch3d = "*"
pytorch_lightning = "*"
hydra-core = "^1.3.2"
code-loader = "^1.0.29"
code-loader-helpers = "^1.0.9"
pyyaml = "^6.0.1"
google-cloud-storage = "^2.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
